# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TAP Backend Laravel 11",
        "description": "Documentación de la API con Sanctum",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/profiles": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Obtener todos los perfiles",
                "operationId": "43d12304c3e621ecce8cc9ccd4b3738e",
                "responses": {
                    "200": {
                        "description": "Respuesta con pefiles",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Almacenar un nuevo perfil",
                "operationId": "00abb5f24a20112b161d17e9a5c38a37",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "sections"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Administrador"
                                    },
                                    "sections": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Users",
                                            "Products",
                                            "Profiles"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil almacenado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profiles/{profile}": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Obtener un perfil",
                "operationId": "83d9b96185bae23c5b0a39be5b9681e7",
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "description": "ULID del perfil",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "ulid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Perfil encontrado",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Perfil no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Actualizar registro de un perfil",
                "operationId": "d287a4ee84dddce8a10e8dccba4c0ddf",
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "description": "ULID del perfil",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "01ARZ3NDEKTSV4RRFFQ69G5FAV"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "sections"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Administrador"
                                    },
                                    "sections": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Users",
                                            "Products",
                                            "Profiles"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil actualizado",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Perfil no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Eliminar registro de un perfil",
                "operationId": "5eb02443a95a4b74b56882493efb5017",
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "description": "ULID del perfil",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "01ARZ3NDEKTSV4RRFFQ69G5FAV"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Perfil eliminado",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Perfil no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profiles/export": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Exportar todos los perfiles",
                "operationId": "31b31191b1fcda948aa85a1e2168f029",
                "responses": {
                    "200": {
                        "description": "Respuesta con usuarios en un xlsx",
                        "content": {
                            "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}